Edited Enemy
Edited enmey class so that now you don't have to keep adding a mediator manually. I also made it so that when you create a new enemy you just create a new inhereted scene of the enemy template scene. In the enemy template scene is where we will add the state machine so that all enemys will be able to use it. We will only add the main state machine howevewr none of the states. The states will be added in the individual enemy scenes